% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_two_rnaseq.R
\name{compare_two_rnaseq}
\alias{compare_two_rnaseq}
\title{compare_two_rnaseq}
\usage{
compare_two_rnaseq(rnaseq_dat, data_1, data_2)
}
\arguments{
\item{rnaseq_dat}{tibble, table of multiple rnaseq data log2 fold-change.
first column should be gene-d, then data and then label (optional)}

\item{data_1}{string, first data to be plotted in comparison}

\item{data_2}{string, second data to be plotted in comparison}
}
\value{
a scatterplot highlighting commonly up- and down- regulated genes
}
\description{
compare_two_rnaseq
}
