% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_qPCR_foldchange.R
\name{compute_qPCR_foldchange}
\alias{compute_qPCR_foldchange}
\title{compute_qPCR_foldchange}
\usage{
compute_qPCR_foldchange(
  dat_qPCR = dat_qPCR,
  house_keeping = "TUB2",
  control = "WT",
  treatment = "WT+drug",
  return_pval = TRUE
)
}
\arguments{
\item{dat_qPCR}{tibble, a tibble of qPCR output. With rows as sample name
ie., WT_R1, WT_R2, WT_R3, treatment_R1,... and columns are genes.
Replicates should always be in the format "_R1" or "_1" or "_set1"}

\item{house_keeping}{character vector, containing name of the house-keeping
control eg. "TUB2" for tubulin gene.}

\item{control}{character vector, determining name of the control samples eg.
WT}

\item{treatment}{character vector, determining name of the treatment samples
eg. treatment}

\item{return_pval}{logical, to return pvalue or tibble of fold-change}
}
\value{
ggplot of the fold change and tibble of foldchange, pvalue
  (determined by t-test)
}
\description{
compute_qPCR_foldchange
}
\details{
pvalue calculation
  source:\url{https://stats.oarc.ucla.edu/other/mult-pkg/faq/pvalue-htm/}
}
\examples{
library(magrittr)
gene_of_interest <- c(22.1, 22.3, 22.4, 8, 8.9, 8.3)
house_keeping_gene <- c(20, 20.1, 20.3,20, 20.4, 19.99)
rname <- c(paste("WT", seq(1:3), sep = "_"), paste("WT+drug", seq(1:3), sep = "_"))

dt_qPCR <- tidyr::tibble(gene_of_interest, house_keeping_gene)
rownames(dt_qPCR) <- rname

dt_qPCR <- dt_qPCR \%>\% tibble::rownames_to_column(var = "Sample")

compute_qPCR_foldchange(dat_qPCR = dt_qPCR, house_keeping = "house_keeping_gene",
control = "WT", treatment = "WT+drug", return_pval=TRUE)

}
