% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pvalue.R
\name{plot_pvalues}
\alias{plot_pvalues}
\title{plot_pvales}
\usage{
plot_pvalues(
  dat_tble,
  display_pval = NULL,
  control_group = "WT",
  scales = "free",
  y_label = "fold_change",
  nrow = 1
)
}
\arguments{
\item{dat_tble}{tibble or data frame with rows=sample and column=conditions (+Pi or -Pi)}

\item{display_pval}{logical, to display pvalue or not. TRUE: Pvalue determined by t-test, FALSE: asterisk, NULL: no pvalue displayed}

\item{control_group}{character vector, which sample to be considered as control to determine pvalue. eg. "WT" *special characters will through error}

\item{scales}{character vector, to keep free scales for each group on y-axis or not. eg. "free" or NULL}

\item{y_label}{character vector, label displayed on y-axis.}

\item{nrow}{numeric, number of rows in which groups should be displayed}
}
\value{
barplot with pvalues displayed
}
\description{
plot_pvales
}
\examples{

library(magrittr)
gene_1 <- c(1.08, 1.01 , 0.917, 0.401, 0.356, 0.450)
gene_2 <- c(1.02, 1.03, 1.002,0.8, 0.76, 0.79)
rname <- c(paste("WT", seq(1:3), sep = "_"), paste("WT+drug", seq(1:3), sep = "_"))
tibble_fc <- tidyr::tibble(gene_1, gene_2)
rownames(tibble_fc) <- rname

tibble_fc <- tibble_fc \%>\% tibble::rownames_to_column(var = "sample")

plot_pvalues(dat_tble = tibble_fc, display_pval = FALSE, control_group = "WT", scales = NULL, y_label = "acid phosphatase activity")

}
