% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qPCR_foldchange.R
\name{plot_qPCR_foldchange}
\alias{plot_qPCR_foldchange}
\title{plot_qPCR_foldchange}
\usage{
plot_qPCR_foldchange(
  fc_tibble,
  display_pval = NULL,
  control_group = "WT",
  nrow = 2
)
}
\arguments{
\item{fc_tibble}{tibble of fold-change values determined as 2^-∆∆Ct, can be
output of compute_qPCR_foldchange function. With rows as sample name ie.,
WT_R1, WT_R2, WT_R3, treatment_R1,... and columns are genes. Replicates
should always be in the format "_R1" or "_1" or "_set1"}

\item{display_pval}{Logical, to display pvalue on the barplot. default: NULL}

\item{control_group}{character vector, sample or group which should be
considered as reference to compute pvalue from t.test.}

\item{nrow}{numeric, number of rows in which the genes should be displayed.
default: 2}
}
\value{
barplot of the fold-change
}
\description{
plot_qPCR_foldchange
}
\details{
pvalue calculation
  source:\url{https://stats.oarc.ucla.edu/other/mult-pkg/faq/pvalue-htm/}
}
\examples{
library(magrittr)
gene_1 <- c(1.08, 1.01 , 0.917, 0.401, 0.356, 0.450)
gene_2 <- c(1.02, 1.03, 1.002,0.8, 0.76, 0.79)
rname <- c(paste("WT", seq(1:3), sep = "_"), paste("WT+drug", seq(1:3), sep = "_"))
qPCR_fc <- tidyr::tibble(gene_1, gene_2)
rownames(qPCR_fc) <- rname

qPCR_fc <- qPCR_fc \%>\% tibble::rownames_to_column(var = "Sample")
plot_qPCR_foldchange(fc_tibble = qPCR_fc, display_pval = FALSE, nrow=1)

}
