% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival_curve.R
\name{plot_survival_curve}
\alias{plot_survival_curve}
\title{plot_sruvival_curve}
\usage{
plot_survival_curve(survival_data, palette = NULL, display_pval = TRUE)
}
\arguments{
\item{survival_data}{status should be 1 or 0, 1=dead, 0=alive}

\item{palette}{character vector to determine a specific color palette or NULL for default palette}

\item{display_pval}{logical, to display pvalue or not}
}
\value{
survival plot
}
\description{
plot_sruvival_curve
}
