% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR_foldchange_heatmap.R
\name{qPCR_foldchange_heatmap}
\alias{qPCR_foldchange_heatmap}
\title{qPCR_foldchange_heatmap}
\usage{
qPCR_foldchange_heatmap(
  fc_tibble,
  control = "WT",
  treatment = "WT+drug",
  threshold = TRUE
)
}
\arguments{
\item{fc_tibble}{tibble of fold-change values determined as 2^-∆∆Ct, can be
output of compute_qPCR_foldchange function. With rows as sample name ie.,
WT_R1, WT_R2, WT_R3, treatment_R1,... and columns are genes. Replicates
should always be in the format "_R1" or "_1" or "_set1"}

\item{control}{character vector, determining name of the control samples eg.
WT}

\item{treatment}{character vector, determining name of the control samples}

\item{threshold}{logical, to limit the color scale of heatmap. default: TRUE
eg. treatment}
}
\value{
heatmap of control and treatment pvalues
}
\description{
qPCR_foldchange_heatmap
}
\details{
pvalue calculation
  source:\url{https://stats.oarc.ucla.edu/other/mult-pkg/faq/pvalue-htm/}
}
\examples{
library(magrittr)
gene_1 <- c(1.08, 1.01 , 0.917, 0.401, 0.356, 0.450)
gene_2 <- c(1.02, 1.03, 1.002,0.8, 0.76, 0.79)
rname <- c(paste("WT", seq(1:3), sep = "_"), paste("WT+drug", seq(1:3), sep = "_"))
qPCR_fc <- tidyr::tibble(gene_1, gene_2)
rownames(qPCR_fc) <- rname

qPCR_fc <- qPCR_fc \%>\% tibble::rownames_to_column(var = "Sample")
qPCR_foldchange_heatmap(fc_tibble = qPCR_fc)

}
